requirements.txt
# AI Agentic Framework Dependencies

# Core dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
PyYAML==6.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Async support
asyncio==3.4.3

# Logging
colorlog==6.8.0

# =====================================
# CONNECTOR DEPENDENCIES
# =====================================

# GitHub API
PyGithub==2.1.1

# Google Drive API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Slack API
slack-sdk==3.26.1

# Email (built-in with Python)
# No additional packages needed for SMTP/IMAP

# =====================================
# OPTIONAL DEPENDENCIES
# =====================================

# NLP and text processing
nltk==3.8.1
spacy==3.7.2

# Machine Learning
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2

# Database support
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# =====================================
# INSTALLATION NOTES
# =====================================
#
# Minimal install (core framework only):
#   pip install aiohttp beautifulsoup4 lxml requests PyYAML pandas numpy
#
# With connectors:
#   pip install -r requirements.txt
#
# For GitHub only:
#   pip install PyGithub requests
#
# For Google Drive only:
#   pip install google-auth google-auth-oauthlib google-api-python-client
#
# For Slack only:
#   pip install slack-sdk requests
#
# Email connector works out of the box (uses Python built-in libraries)
